requirements.txt
flask==1.1.2
line-bot-sdk==1.16.0
